Move ´get_axis´ function to ´data_loader´
